Conflicts:
	rts/Sim/Units/COB/CobInstance.cpp
	rts/Sim/Units/CommandAI/TransportCAI.cpp
	rts/Sim/Units/Unit.cpp
	rts/Sim/Units/UnitLoader.cpp
	rts/Sim/Units/UnitTypes/Builder.cpp
	rts/Sim/Units/UnitTypes/Factory.cpp
	rts/Sim/Weapons/BeamLaser.cpp

+++ rts/Map/SM3/terrain/Terrain.cpp
@@ -748,7 +748,7 @@ namespace terrain {
 
 		texturing->BeginTexturing();
 		// render to the framebuffer
-		for (int p=0;p<q->textureSetup->renderSetup [0]->passes.size();p++)
+		for (size_t p=0; p < q->textureSetup->renderSetup[0]->passes.size(); p++)
 		{
 			texturing->BeginPass(p);
 			if (texturing->SetupNode (q, p))
